requirements.txt
openai
langchain-community
streamlit
python-dotenv
PyPDF2


-e .

